Fix: WBS names now display on startup
The issue was that the WBSVM that was referenced by the Task VMs were not
the same objects as those in the WBSVM list because of how I was looping through them.
The SelectedItem needs to be an object with the same reference as an item in
the ItemsSource of a combo box.